Translators fix

+++ src/test/java/com/google/security/zynamics/reil/translators/ppc/AddiTranslatorTest.java
@@ -62,7 +62,7 @@ public class AddiTranslatorTest {
 
   @Test
   public void testSimple() throws InternalTranslationException, InterpreterException {
-    interpreter.setRegister("%r1", BigInteger.valueOf(0x80000000L), OperandSize.BYTE,
+    interpreter.setRegister("%r1", BigInteger.valueOf(0x80000000L), OperandSize.DWORD,
         ReilRegisterStatus.DEFINED);
     interpreter.setRegister("%r0", BigInteger.ZERO, OperandSize.DWORD, ReilRegisterStatus.DEFINED);
