+++ src/test/java/com/google/security/zynamics/reil/translators/ppc/AddicTranslatorTest.java
@@ -65,7 +65,7 @@ public class AddicTranslatorTest {
   public void testCarryPropagate() throws InternalTranslationException, InterpreterException {
     interpreter
         .setRegister("XERCA", BigInteger.ZERO, OperandSize.DWORD, ReilRegisterStatus.DEFINED);
-    interpreter.setRegister("%r1", BigInteger.valueOf(0x80000000L), OperandSize.BYTE,
+    interpreter.setRegister("%r1", BigInteger.valueOf(0x80000000L), OperandSize.DWORD,
         ReilRegisterStatus.DEFINED);
     interpreter.setRegister("%r0", BigInteger.ZERO, OperandSize.DWORD, ReilRegisterStatus.DEFINED);
 
@@ -101,7 +101,7 @@ public class AddicTranslatorTest {
   public void testSimple() throws InternalTranslationException, InterpreterException {
     interpreter
         .setRegister("XERCA", BigInteger.ZERO, OperandSize.DWORD, ReilRegisterStatus.DEFINED);
-    interpreter.setRegister("%r1", BigInteger.valueOf(0x80000000L), OperandSize.BYTE,
+    interpreter.setRegister("%r1", BigInteger.valueOf(0x80000000L), OperandSize.DWORD,
         ReilRegisterStatus.DEFINED);
     interpreter.setRegister("%r0", BigInteger.ZERO, OperandSize.DWORD, ReilRegisterStatus.DEFINED);
 
@@ -137,7 +137,7 @@ public class AddicTranslatorTest {
   public void testSingExtend() throws InternalTranslationException, InterpreterException {
     interpreter
         .setRegister("XERCA", BigInteger.ZERO, OperandSize.DWORD, ReilRegisterStatus.DEFINED);
-    interpreter.setRegister("%r1", BigInteger.ONE, OperandSize.BYTE, ReilRegisterStatus.DEFINED);
+    interpreter.setRegister("%r1", BigInteger.ONE, OperandSize.DWORD, ReilRegisterStatus.DEFINED);
     interpreter.setRegister("%r0", BigInteger.ZERO, OperandSize.DWORD, ReilRegisterStatus.DEFINED);
 
     final MockOperandTree operandTree1 = new MockOperandTree();
